Adding AiiDA and improving description of Materials Cloud
Similar to #143, where `httk` was added for the software
in parallel to `omdb` for the database, here we add `aiida`
as the prefix for the AiiDA software (www.aiida.net) in parallel
to the `mcloud` prefix for the Materials Cloud (www.materialscloud.org)
online database.

Moreover, I'm clarifying and improving the description of the
Materials Cloud entry.