Define returning properties with unknown values
REQUIRED properties with an unknown value MUST be returned (unless it is
left out through the use, e.g., `response_fields`).
OPTIONAL properties with an unknown value MAY be returned, but if they
left out in a _full_ response, the client MUST assume they have an
unknown value, i.e., is `null`.