Change unit definitions so the way we reference the standard from schemas make sense. Important: this changes how units are ordered in Compound Unit Expressions
Allowing prefixes to be merged with unit symbols greatly complicates parsing and may not even make sense (or be unambigous) the way we now can mix and match symbols.
From a standards and implementation prespective it is also preferably not to have to treat prefixes differently from unitless constants.
Now all unitless prefixes and constants go first (in alphabetical order) and then comes all "real" units (in alphabetical order).